# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API",
    "version": "1.0.0"
  },
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "modelo": {
          "type": "string"
        },
        "cv": {
          "type": "number"
        },
        "precio": {
          "type": "number"
        }
      },
      "required": ["modelo", "cv", "precio"]
    },
    "Concesionario": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "coches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coche"
          }
        }
      },
      "required": ["nombre", "direccion"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Obtener todos los concesionarios",
        "operationId": "getConcesionarios",
        "responses": {
          "200": {
            "description": "Lista de concesionarios",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionario"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo concesionario",
        "operationId": "createConcesionario",
        "parameters": [
          {
            "in": "body",
            "name": "concesionario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario creado"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Obtener un concesionario por ID",
        "operationId": "getConcesionarioById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del concesionario",
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          },
          "404": {
            "description": "Concesionario no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un concesionario por ID",
        "operationId": "updateConcesionarioById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "concesionario",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Concesionario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario actualizado"
          },
          "404": {
            "description": "Concesionario no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Borrar un concesionario por ID",
        "operationId": "deleteConcesionarioById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Concesionario borrado"
          },
          "404": {
            "description": "Concesionario no encontrado"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Obtener todos los coches de un concesionario",
        "operationId": "getCochesByConcesionarioId",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de coches",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Coche"
              }
            }
          },
          "404": {
            "description": "Concesionario no encontrado"
          }
        }
      },
      "post": {
        "summary": "Añadir un coche a un concesionario",
        "operationId": "addCocheToConcesionario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "coche",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche añadido"
          },
          "404": {
            "description": "Concesionario no encontrado"
          }
        }
      }
    },
    "/concesionarios/{id}/coches/{cocheId}": {
      "get": {
        "summary": "Obtener un coche por ID de un concesionario",
        "operationId": "getCocheById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del coche",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          },
          "404": {
            "description": "Coche no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un coche por ID",
        "operationId": "updateCocheById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "coche",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche actualizado"
          },
          "404": {
            "description": "Coche no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un coche por ID",
        "operationId": "deleteCocheById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "cocheId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Coche eliminado"
          },
          "404": {
            "description": "Coche no encontrado"
          }
        }
      }
    }
  }
}